refactor: move BaseContract to own file
